requirements.txt
numpy
ttkbootstrap
matplotlib
pyserial